+++ package.json
@@ -8,6 +8,7 @@
     "url": "https://github.com/yahoo/react-cartographer"
   },
   "scripts": {
+    "compile": "babel src --out-dir lib",
     "lint": "jshint",
     "start": "npm run serve | npm run dev",
     "serve": "./node_modules/.bin/http-server -p 8080",
@@ -15,7 +16,7 @@
     "test": "jenkins-mocha --compilers js:babel/register tests/unit/*",
     "devtest": "mocha --compilers js:babel/register tests/unit/* --reporter nyan",
     "cover": "node node_modules/istanbul/lib/cli.js cover --dir artifacts -- ./node_modules/mocha/bin/_mocha tests/ --recursive --reporter spec",
-    "prepublish": "grunt babel"
+    "prepublish": "npm run compile"
   },
   "dependencies": {
     "url": "^0.10.0",
