+++ src/lib/bingMap.js
@@ -6,8 +6,8 @@
 'use strict';
 
 /*
- * Yahoo MapImage Service
- * @class GoogleMapService
+ * Bing MapImage Service
+ * @class BingMapService
  */
 
 const _ = {
@@ -16,54 +16,52 @@ const _ = {
     identity: require('lodash/utility/identity')
 };
 
-import {yahoo as config} from '../config/config.json';
+import {bing as config} from '../../config/config.json';
 import utils from 'url';
+import encode from 'urlencode';
 
-export default class YahooMapService {
+export default class BingMapService {
     get name() {
         return config.name;
     }
 
     /**
      * Get the map location details for the address provided in the params
-     * Map Location details include: geo coordinates and map url to display
      *
      * @param {Object} params
      * @returns {{mapId: (*|string|string|string|string), data: {locationLink: *, locationText: string}}}
      */
     getMap (params) {
+        let locationText;
         let location;
+        let pushpin;
         let url;
-        const query = _.pick({
-            appid: params.providerKey,
-            imw: params.width,
-            imh: params.height,
-            imi: config.imi,
-            radius: config.radius,
-            zoom: params.zoom
-        }, _.identity);
 
         if (isFinite(params.longitude) && isFinite(params.latitude)) {
-            location = [params.latitude, params.longitude].join(',');
-            query.clat = params.latitude;
-            query.clon = params.longitude;
+            pushpin = location = locationText = [params.latitude, params.longitude].join(',');
+            if (isFinite(params.zoom)) {
+                location = [location, params.zoom].join('/');
+            }
         } else {
-            location = [params.line1, params.line2, params.line3].join(',');
-            query.q = location;
+            location = locationText = [params.line1, params.line2, params.line3].join(',');
         }
 
         url = utils.format({
             protocol: config.protocol,
             hostname: config.host,
-            pathname: config.path,
-            query: query
+            pathname: config.path + encode(location),
+            query: _.pick({
+                mapSize: [params.width, params.height].join(','),
+                key: params.providerKey,
+                pushpin: pushpin
+            }, _.identity)
         });
 
         return {
             mapId: params.mapId,
             data: {
                 locationLink: url,
-                locationText: location
+                locationText: locationText
             }
         };
     }
