+++ src/lib/mapLocationFactory.js
